requirements.txt
aiohttp==3.7.4.post0
argcomplete==1.12.3
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
colorclass==2.2.0
decorator==5.1.0
idna==3.2
macaroonbakery==1.3.1
multidict==5.2.0
oauthlib==3.1.1
protobuf==3.18.0
pycparser==2.20
pymacaroons==0.13.0
pymongo==3.12.0
PyNaCl==1.4.0
pyRFC3339==1.1
python-libmaas==0.6.6
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
terminaltables==3.1.0
typing-extensions==3.10.0.2
urllib3==1.26.7
validators==0.18.2
yarl==1.6.3
